+++ server/routes/authRoutes.js
@@ -1,15 +1,8 @@
 const express = require("express");
 const router = express.Router();
-const { registerUser, loginUser } = require("../controllers/authController");
+const { loginUser, logoutUser } = require("../controllers/authController");
 
-router.post("/register", registerUser);
 router.post("/login", loginUser);
-router.get("/", (req, res) => {
-    console.log(req.cookies);
-    res.json({
-        get: "jajal token",
-        cookie: req.cookies
-    });
-})
+router.post("/logout", logoutUser);
 
 module.exports = router;
